requirements.txt
python-dotenv==0.10.1
expiringdict==1.1.4
-e git+https://github.com/Rapptz/discord.py@rewrite#egg=discordpy
romkan==0.2.1
selenium==3.141.0
Pillow==6.0.0
